requirements.txt
selenium
pandas
openpyxl
pillow
undetected-chromedriver